requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
argcomplete==3.6.2
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
Bootstrap-Flask==2.5.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.12
click==8.3.1
colorama==0.4.6
Flask==3.1.2
Flask-CKEditor==0.4.6
Flask-Cors==5.0.0
Flask-Gravatar==0.5.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
frozenlist==1.6.0
greenlet==3.2.3
idna==3.2
itsdangerous==2.2.0
Jinja2==3.1.6
mailtrap==2.0.1
MarkupSafe==3.0.3
multidict==6.4.3
numpy==2.2.4
packaging==25.0
pandas==2.2.3
pipx==1.7.1
platformdirs==4.4.0
propcache==0.3.1
PyJWT==2.10.1
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.26.0
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.25
stripe==12.0.0
toml==0.10.2
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
userpath==1.9.2
Werkzeug==3.1.5
WTForms==3.0.1
yarl==1.20.0
